make this portable across multiple OS's (don't assume root access); fix cluster so load-balancer gets updated; and publish the URL
there is also a needed fix in brooklyncentral/brooklyn#409 which corrects the way Host is set when we proxypass
(without this, images don't render in clustered drupal)

finally -- it seems like email isn't working? so we can't actually log in and get started.
would be good to configure email and/or create an initial user.